requirements.txt
Flask==2.3.2
redis==5.0.0
python-dotenv==1.0.0